fix swapping, changed from Double to BigDecimal

+++ src/main/kotlin/co/akoot/plugins/bluefox/api/economy/WorldWallet.kt
@@ -1,7 +1,9 @@
 package co.akoot.plugins.bluefox.api.economy
 
+import java.math.BigDecimal
+
 object WorldWallet: Wallet(1, "WORLD") {
-    override fun send(wallet: Wallet, coin: Coin, amount: Double, relatedId: Int?): Int {
+    override fun send(wallet: Wallet, coin: Coin, amount: BigDecimal, relatedId: Int?): Int {
         balance[coin] = amount // nice hacks
         return super.send(wallet, coin, amount, relatedId)
     }
